Primeira entrega

+++ src/components/BoxEditUser/index.tsx
@@ -12,15 +12,18 @@ import { RoutePath } from 'types/routes';
 import { updateService } from 'services/updateService';
 import {
 	alertaDelete,
-	alertaDeleteOtherUser,
+	alertaErro,
 	alertaSucesso,
+	alertaUpdate,
 } from 'utils/alertas';
+import ModalPassword from 'components/ModalPassword';
 
 const BoxEditUser = () => {
 	const navigate = useNavigate();
 	const userIdStorage = localStorage.getItem('userIdStorage');
 	const [users, setUsers] = useState<userObj[]>([]);
 	const [refreshUsers, setRefreshUsers] = useState(false);
+	const [isModalOpen, setIsModalOpen] = useState(false);
 	const [userId, setUserId] = useState({
 		id: '',
 	});
@@ -74,6 +77,8 @@ const BoxEditUser = () => {
 		// eslint-disable-next-line react-hooks/exhaustive-deps
 	}, [userId, refreshUsers]);
 
+	const admin = users.filter((e) => e.name === 'Administrador');
+
 	const updateUser = async () => {
 		console.log(userId.id, user);
 		const values = {
@@ -81,18 +86,23 @@ const BoxEditUser = () => {
 			email: user.email,
 			isAdmin: user.isAdmin,
 		};
-		const response = await updateService.updateUser2(userId.id, values);
-		console.log(response);
 
-		if (response.status === 200) {
-			alertaSucesso.alerta('Usuario Atualizado com sucesso!');
-			updateUsers(true);
+		if (admin[0].id !== userIdStorage) {
+			alertaUpdate.updateUser();
+		} else {
+			const response = await updateService.updateUserEdit(userId.id, values);
+			console.log(response);
+
+			if (response.status === 200) {
+				alertaSucesso.alerta('Usuario Atualizado com sucesso!');
+				updateUsers(true);
+			}
 		}
 	};
 
 	const deleteModalOpen = () => {
-		if (userId.id !== userIdStorage) {
-			alertaDeleteOtherUser();
+		if (admin[0].id !== userIdStorage) {
+			alertaDelete.deleteAdmin();
 		} else {
 			alertaDelete.deleteUser().then((resp) => {
 				console.log(resp);
@@ -104,13 +114,23 @@ const BoxEditUser = () => {
 	};
 
 	const deleteUser = async () => {
-		const response = await deleteService.deleteUser(`${userId.id}`);
-		console.log(response);
-		alertaSucesso.alerta('Usuario deletado com sucesso!');
-		localStorage.removeItem('jwtLocalStorage');
-		localStorage.removeItem('userIdStorage');
-		localStorage.removeItem('profileId');
-		navigate(RoutePath.LOGIN);
+		if (admin[0].id === userIdStorage) {
+			if (userId.id !== userIdStorage) {
+				const response = await deleteService.deleteUser(`${userId.id}`);
+				console.log(response);
+				alertaSucesso.alerta('Usuario deletado com sucesso!');
+			} else {
+				alertaErro.alerta('Esse usuario não pode ser deletado!');
+			}
+		} else {
+			const response = await deleteService.deleteUser(`${userId.id}`);
+			console.log(response);
+			alertaSucesso.alerta('Usuario deletado com sucesso!');
+			localStorage.removeItem('jwtLocalStorage');
+			localStorage.removeItem('userIdStorage');
+			localStorage.removeItem('profileId');
+			navigate(RoutePath.LOGIN);
+		}
 	};
 
 	const updateUsers = (refresh: boolean) => {
@@ -120,67 +140,87 @@ const BoxEditUser = () => {
 		}, 100);
 	};
 
+	const openModal = () => {
+		setIsModalOpen(true);
+	};
+
+	function closeModal() {
+		setIsModalOpen(false);
+	}
+
 	return (
-		<S.BoxEditUser>
-			<S.BoxEditUserLogo>
-				<S.BoxEditUserLogoImage
-					src={logo}
-					alt="Image do logo com um desenho de um controle de video game"
-				/>
-			</S.BoxEditUserLogo>
-			<S.BoxEditUserForm>
-				<S.BoxEditUserSearch>
-					<select onChange={handleChangeOption} name="name" id="name">
-						<optgroup label="Usuarios">
-							<option>Escolha</option>
-							{users.map((user, index) => (
-								<option value={user.id} key={index}>
-									{user.name}
-								</option>
-							))}
-						</optgroup>
-					</select>
-				</S.BoxEditUserSearch>
-				<input
-					type="text"
-					name="name"
-					id="name"
-					placeholder="Nome..."
-					onChange={handleChangeValues}
-					defaultValue={user.name}
-				/>
-				<input
-					type="email"
-					name="email"
-					id="email"
-					placeholder="Email..."
-					onChange={handleChangeValues}
-					defaultValue={user.email}
-				/>
-				<S.BoxEditUserSearch>
-					<select onChange={handleChangeOption2} name="isAdmin" id="isAdmin">
-						<optgroup label="Tipo">
-							<option value="true">Admin</option>
-							<option value="false">User</option>
-						</optgroup>
-					</select>
-				</S.BoxEditUserSearch>
-				<ButtonResetSenha value="Resetar Senha" type="button" id="resetSenha" />
-				<div id="buttons">
-					<ButtonDeletar
-						value="Deletar"
-						type="button"
-						onClick={deleteModalOpen}
+		<>
+			<S.BoxEditUser>
+				<S.BoxEditUserLogo>
+					<S.BoxEditUserLogoImage
+						src={logo}
+						alt="Image do logo com um desenho de um controle de video game"
 					/>
-					<ButtonAtualizar
-						value="Atualizar"
-						type="button"
-						onClick={updateUser}
+				</S.BoxEditUserLogo>
+				<S.BoxEditUserForm>
+					<S.BoxEditUserSearch>
+						<select onChange={handleChangeOption} name="name" id="nameSelect">
+							<optgroup label="Usuarios">
+								<option>Escolha</option>
+								{users.map((user, index) => (
+									<option value={user.id} key={index}>
+										{user.name}
+									</option>
+								))}
+							</optgroup>
+						</select>
+					</S.BoxEditUserSearch>
+					<input
+						type="text"
+						name="name"
+						id="name"
+						placeholder="Nome..."
+						onChange={handleChangeValues}
+						defaultValue={user.name}
 					/>
-				</div>
-			</S.BoxEditUserForm>
-		</S.BoxEditUser>
+					<input
+						type="email"
+						name="email"
+						id="email"
+						placeholder="Email..."
+						onChange={handleChangeValues}
+						defaultValue={user.email}
+					/>
+					<S.BoxEditUserSearch>
+						<select onChange={handleChangeOption2} name="isAdmin" id="isAdmin">
+							<optgroup label="Tipo">
+								<option value="false">User</option>
+								<option value="true">Admin</option>
+							</optgroup>
+						</select>
+					</S.BoxEditUserSearch>
+					<ButtonResetSenha
+						value="Resetar Senha"
+						type="button"
+						id="resetSenha"
+						onClick={openModal}
+					/>
+					<div id="buttons">
+						<ButtonDeletar
+							value="Deletar"
+							type="button"
+							onClick={deleteModalOpen}
+						/>
+						<ButtonAtualizar
+							value="Atualizar"
+							type="button"
+							onClick={updateUser}
+						/>
+					</div>
+				</S.BoxEditUserForm>
+			</S.BoxEditUser>
+			<ModalPassword
+				isOpen={isModalOpen}
+				closeModal={closeModal}
+				userId={userId.id}
+			/>
+		</>
 	);
-};
+};;
 
 export default BoxEditUser;
